refactor: use existing helper for display name
