+++ src/components/Contractor/Payments/CreatePayment/PreviewPresentation.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react'
 import type { Contractor } from '@gusto/embedded-api/models/components/contractor'
 import type { CompanyBankAccount } from '@gusto/embedded-api/models/components/companybankaccount'
 import type { PaymentSpeed } from '@gusto/embedded-api/models/components/paymentconfigs'
-import { getContractorDisplayName } from './helpers'
+import { getContractorDisplayName } from '../../shared/helpers'
 import { FastAchSubmissionBlockerBanner } from './FastAchSubmissionBlockerBanner'
 import { GenericBlocker } from './GenericBlocker'
 import { DataView, Flex } from '@/components/Common'
