+++ src/components/Contractor/Payments/PaymentHistory/PaymentHistoryPresentation.tsx
@@ -2,7 +2,7 @@ import { Trans, useTranslation } from 'react-i18next'
 import type { ContractorPaymentGroup } from '@gusto/embedded-api/models/components/contractorpaymentgroup'
 import type { ContractorPaymentForGroup } from '@gusto/embedded-api/models/components/contractorpaymentforgroup'
 import type { Contractor } from '@gusto/embedded-api/models/components/contractor'
-import { getContractorDisplayName } from '../CreatePayment/helpers'
+import { getContractorDisplayName } from '../../shared/helpers'
 import styles from './PaymentHistoryPresentation.module.scss'
 import { DataView, Flex, EmptyData } from '@/components/Common'
 import { useComponentContext } from '@/contexts/ComponentAdapter/useComponentContext'
