+++ src/components/Contractor/Payments/PaymentSummary/PaymentSummaryPresentation.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react'
 import type { Contractor } from '@gusto/embedded-api/models/components/contractor'
 import type { CompanyBankAccount } from '@gusto/embedded-api/models/components/companybankaccount'
 import type { InternalAlert } from '../types'
-import { getContractorDisplayName } from '../CreatePayment/helpers'
+import { getContractorDisplayName } from '../../shared/helpers'
 import { DataView, Flex } from '@/components/Common'
 import { useComponentContext } from '@/contexts/ComponentAdapter/useComponentContext'
 import { useI18n } from '@/i18n'
